Map layers sync with map button state on initialization
Was trying to reproduce the problem where filter buttons are out of sync
with map layers. Couldn't reproduce, but added code to check whether the
filter button has an 'active' class and uses that to determine whether
or not to show the layer on initialization.
Additionally, icons/text inside the button have classes that are
supposed to be in sync with the button's 'active' class. To make sure
they never get out of sync, rather than toggling all classes, the
button's state gets toggled, and then all others set their classes
depending on the 'active' status of the button.
These are all small things that probably aren't what was causing the
actual bug. But there's a chance that they'll catch it