Unifies how spill/complaint date ranges are calculated
window.complaint_timeframe and window.spill_timeframe objects added that
have number of days to sample across, the last time the data was
updated, and functions to get sample dates.
Sidebar, map, and pills ontop of map all use this now to ensure they all
use the same way of calculating the timeframe of the data shown.
Also updated labels from "in the last x days" to "from Sep 10 to Aug 10"